requirements.txt
click==8.0.1
cycler==0.10.0
Flask==2.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
